Refactor JWT auth and middleware, update access levels
Removed JwtAuthorizeAttribute class and attribute usage.
Changed ADMIN_TEMPLATE_PHYSICAL_PATH to public.
Added JwtAuthorizationMiddleware for JWT processing.
Reintroduced ApplicationExecutionContextMiddleware with docs.
