--- src/Backend/FluentCMS.Web.Api/Attributes/JwtAuthorizeAttribute.cs
@@ -1,13 +1,0 @@
-﻿using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.AspNetCore.Authorization;
-
-namespace FluentCMS.Web.Api;
-
-[AttributeUsage(AttributeTargets.Class | AttributeTargets.Method, AllowMultiple = false, Inherited = true)]
-public class JwtAuthorizeAttribute : AuthorizeAttribute
-{
-    public JwtAuthorizeAttribute()
-    {
-        AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme;
-    }
-}
