+++ src/Backend/FluentCMS.Web.Api/Controllers/BaseController.cs
@@ -4,7 +4,6 @@ namespace FluentCMS.Web.Api.Controllers;
 
 [ApiController]
 [Produces("application/json")]
-[JwtAuthorize]
 [TypeFilter(typeof(ApiTokenAuthorizeFilter))]
 public abstract class BaseController
 {
